fill in variable in defined commands
The game engine can now understand <?> as meaning "fill in what the player wrote here. By adding this to both the command that will be written by the player, and adding it in the command, the game engine will fill in the correct text in the correct place for the command.