Multichoice, footnotes, and roomCommands
Added functionality that allows for multichoice stories with the help of footnotes declared with a # that is used by the game Engine to know where in the file you want the game to continue from.

Also added room specific commands, letting you have commands that only work when you are in specific rooms. These commands are handled the exact same way as normal commands by the game Engine