Added investigate
Created a new way of interacting with items in a room. This function also comes with the new "investigateables" attribute added to rooms. Investigateables have a key value that is used as their name, and include a list of commands in the .ta format that gets run by the game engine. This lets you make custom events for investigating items in a room